+++ src/App.js
@@ -1,123 +1,19 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- *
- * @format
- * @flow strict-local
- */
+// Punto de entrada a la aplicación, donde configuramos los providers, wrappers
+// y demás.
+import React from 'react'
+import { NavigationContainer } from '@react-navigation/native'
 
-import React, { useState } from 'react'
-import type { Node } from 'react'
-import database from '@react-native-firebase/database'
-import {
-  SafeAreaView,
-  ScrollView,
-  StatusBar,
-  StyleSheet,
-  Text,
-  useColorScheme,
-  View
-} from 'react-native'
+import MainNavigation from './MainNavigation'
 
-import {
-  Colors,
-  DebugInstructions,
-  Header,
-  LearnMoreLinks,
-  ReloadInstructions
-} from 'react-native/Libraries/NewAppScreen'
-
-const Section = ({ children, title }): Node => {
-  const isDarkMode = useColorScheme() === 'dark'
-
+const App = () => {
+  // El contenedor maneja el estado de la navegación y se encarga de cosas como
+  // el deep linking y el botón de volver en Android.
   return (
-    <View style={styles.sectionContainer}>
-      <Text
-        style={[
-          styles.sectionTitle,
-          {
-            color: isDarkMode ? Colors.white : Colors.black
-          }
-        ]}>
-        {title}
-      </Text>
-      <Text
-        style={[
-          styles.sectionDescription,
-          {
-            color: isDarkMode ? Colors.light : Colors.dark
-          }
-        ]}>
-        {children}
-      </Text>
-    </View>
+    <NavigationContainer>
+      { /* El navegador principal que gestiona las diferentes pantallas. */ }
+      <MainNavigation />
+    </NavigationContainer>
   )
 }
-
-const App: () => Node = () => {
-  const [gato, setGato] = useState('Estado de un gato')
-
-  const isDarkMode = useColorScheme() === 'dark'
-
-  const backgroundStyle = {
-    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter
-  }
-
-  database().ref('/gatxs/gato11').on('value', snapshot => {
-    setGato(snapshot.val())
-  })
-
-  return (
-    <SafeAreaView style={backgroundStyle}>
-      <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
-      <ScrollView
-        contentInsetAdjustmentBehavior="automatic"
-        style={backgroundStyle}>
-        <Header />
-        <View
-          style={{
-            backgroundColor: isDarkMode ? Colors.black : Colors.white
-          }}>
-          <Section title="TenderCats">
-            Edit <Text style={styles.highlight}>App.js</Text> to change this
-            screen and then come back to see your edits.
-          </Section>
-          <Section title="Prueba de conexión a DB">
-            <Text style={styles.highlight}>{ gato }</Text>
-          </Section>
-          <Section title="See Your Changes">
-            <ReloadInstructions />
-          </Section>
-          <Section title="Debug">
-            <DebugInstructions />
-          </Section>
-          <Section title="Learn More">
-            Read the docs to discover what to do next:
-          </Section>
-          <LearnMoreLinks />
-        </View>
-      </ScrollView>
-    </SafeAreaView>
-  )
-}
-
-const styles = StyleSheet.create({
-  sectionContainer: {
-    marginTop: 32,
-    paddingHorizontal: 24
-  },
-  sectionTitle: {
-    fontSize: 24,
-    fontWeight: '600'
-  },
-  sectionDescription: {
-    marginTop: 8,
-    fontSize: 18,
-    fontWeight: '400'
-  },
-  highlight: {
-    fontWeight: '700'
-  }
-})
 
 export default App
